+++ packages/bluesky/lib/src/services/types/feed/defs/_z.dart
@@ -23,6 +23,7 @@ export 'extensions/generator_view_extension.dart';
 export 'extensions/not_found_post_extension.dart';
 export 'extensions/post_view_extension.dart';
 export 'extensions/viewer_state_extension.dart';
+export 'extensions/post_view_embed_extension.dart';
 
 export 'unions/feed_view_post_reason.dart';
 export 'unions/post_view_embed.dart';