+++ packages/bluesky/lib/src/services/types/feed/defs/extensions/post_view_embed_extension.dart
@@ -1,0 +1,12 @@
+// Copyright 2024 Shinya Kato. All rights reserved.
+// Redistribution and use in source and binary forms, with or without
+// modification, are permitted provided the conditions.
+
+import '../../../embed/record/unions/view_record_embeds.dart';
+import '../unions/post_view_embed.dart';
+
+extension UFeedDefsPostViewEmbedExtension on UFeedDefsPostViewEmbed {
+  /// Returns this object as [UEmbedRecordViewRecordEmbeds].
+  UEmbedRecordViewRecordEmbeds asUEmbedRecordViewRecordEmbeds() =>
+      unionEmbedRecordViewRecordEmbedsConverter.fromJson(toJson());
+}